+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Added
 - 'ServerMetadata' endpoint to the API (:commit:`7747911`)
 - 'modelList' endpoint to the API (:commit:`7477b7d`)
 - Parse JSON data as string in HTTP body (:commit:`694800e`)
+- Directory monitoring for model loading (:commit:`6459797`)
 
 Changed
 ^^^^^^^